Screenshot generation compression
When using colab to generate screenshots, saving them would become a massive pain since the machine was compressing some 80,000 files. Now screenshot_generator saves game frames as numpy arrays using the savez_npz numpy function. Frames are saved in batches of 100 to reduce the number of files that require compressing at the end of the generation.