Tweaking screenshotter parameters
We are still undecided on which world model to use and some of them require video training data while some others require only still images. In the case of still image training data it's okay to only save every 5th frame for example, but in the case of video training data the image stream has to be more consistent. So now the screenshotter will save every frame and has had the number of bots it runs reduced to only 10 to increase each bot's framerate.