Fixed screenshotter hanging on client death
When a client requests to respawn the server used to only send back playerInfo, meaning the client would never send another move and continue the tick loop. Now the server sends both playerInfo and gameSetup, so the client effectively reconnects to the server as if they were joining for the first time.