Screenshot generator utility functions
You can now set the number of bots to join a server along with the number of bots that should capture their game video feeds. Tweaked the server further to feel more like the real game (again). Added a spawn_bots.py script which does exactly what it says on the tin, spawns some dumb ai to join the server.